requirements.txt
selenium
pandas
openpyxl
pillow